requirements.txt
pymongo
langchain
langsmith
python-dotenv